// src/app/recipe/model/recipe-search/recipe-search-filter-key.enum.ts
export enum RecipeSearchFilterKey {
  USER_RECIPES = "userRecipes",
  FAVORITE_RECIPES = "favoriteRecipes",

  RECIPE_AUTHOR = "author",

  RECIPE_TAGS = "tags",
  RECIPE_NAME = "name",
  RECIPE_DESCRIPTION = "description",
  RECIPE_SERVINGS = "servings",
  RECIPE_PREPARATION_TIME = "preparationTime",
  RECIPE_COOKING_TIME = "cookingTime",
  RECIPE_INGREDIENTS = "ingredients",

  RECIPE_AVERAGE_RATING = "averageRating",
}